allow Numeric for iat field on OIDC tokens

+++ app/forms/openid_connect_token_form.rb
@@ -128,7 +128,7 @@ class OpenidConnectTokenForm
   def validate_iat(payload)
     return true unless payload.key?('iat')
     iat = payload['iat']
-    return true if iat.is_a?(Integer) && (iat.to_i - ISSUED_AT_LEEWAY_SECONDS) < Time.zone.now.to_i
+    return true if iat.is_a?(Numeric) && (iat.to_i - ISSUED_AT_LEEWAY_SECONDS) < Time.zone.now.to_i
 
     errors.add(:client_assertion, t('openid_connect.token.errors.invalid_iat'))
   end
